requirements.txt
hubspot-api-client>=8.0.0
snowflake-connector-python>=3.12.0
python-dotenv>=1.0.0
dbt-core>=1.8.0
dbt-snowflake>=1.8.0
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
firecrawl-py>=1.0.0
